tutorials/compare_standards/python/requirements.txt
# HTTP requests
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0
